+++ src/components/layout/index.ts
@@ -1,4 +1,5 @@
 export { AppHeader } from "./AppHeader";
 export { AppShell } from "./AppShell";
 export { BootScreen } from "./BootScreen";
+export { OverlayScrollArea } from "./OverlayScrollArea";
 export { OverlayScrollbar } from "./OverlayScrollbar";